+++ src/base_tools.rs
@@ -57,3 +57,4 @@ pub fn parse_with_base(s: String, p: Parsing) -> Result<usize, ()> {
         },
     }
 }
+
